# Cognitive Execution Layer (CEL)/docs/api/openapi-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cognitive Execution Layer (CEL) API",
    "description": "Autonomous Engineering API for AI-powered code development and system management",
    "version": "1.0.0",
    "contact": {
      "name": "CEL Development Team",
      "email": "support@cel-project.org"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.cel-project.org",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/v1/chat/completions": {
      "post": {
        "summary": "Chat completions endpoint",
        "description": "Generate AI responses for chat interactions",
        "tags": ["AI"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCompletionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat completion generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatCompletionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid input"
          }
        }
      }
    },
...
    "/v1/code-assist": {
      "post": {
        "summary": "Get AI code assistance",
        "description": "Get AI-powered suggestions and assistance for code development",
        "tags": ["AI"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeAssistRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Code assistance provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodeAssistResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatCompletionRequest": {
        "type": "object",
        "required": ["messages"],
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "role": {
                  "type": "string",
                  "enum": ["system", "user", "assistant"],
                  "example": "user"
                },
                "content": {
                  "type": "string",
                  "example": "Write a function to calculate fibonacci numbers"
                }
              }
            }
          },
          "model": {
            "type": "string",
            "example": "gpt-4-turbo"
          },
          "temperature": {
            "type": "number",
            "minimum": 0,
            "maximum": 2,
            "example": 0.7
          }
        }
      },
      "ChatCompletionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "chatcmpl-123"
          },
          "choices": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "object",
                  "properties": {
                    "role": {
                      "type": "string",
                      "example": "assistant"
                    },
                    "content": {
                      "type": "string",
                      "example": "Here's a function to calculate fibonacci numbers..."
                    }
                  }
                },
                "finish_reason": {
                  "type": "string",
                  "example": "stop"
                }
              }
            }
          },
          "usage": {
            "type": "object",
            "properties": {
              "prompt_tokens": {
                "type": "number",
                "example": 50
              },
              "completion_tokens": {
                "type": "number",
                "example": 120
              },
              "total_tokens": {
                "type": "number",
                "example": 170
              }
            }
          }
        }
      },
...
      "CodeAssistRequest": {
        "type": "object",
        "required": ["task"],
        "properties": {
          "task": {
            "type": "string",
            "example": "Refactor this function to be more efficient"
          },
          "code": {
            "type": "string",
            "example": "function inefficientFunction(arr) { ... }"
          },
          "context": {
            "type": "string",
            "example": "Performance optimization request"
          }
        }
      },
      "CodeAssistResponse": {
        "type": "object",
        "properties": {
          "suggestions": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "refactor"
                },
                "description": {
                  "type": "string",
                  "example": "Optimize loop structure"
                },
                "code": {
                  "type": "string",
                  "example": "const optimizedCode = ..."
                }
              }
            }
          },
          "confidence": {
            "type": "number",
            "example": 0.95
          }
        }
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
}